fix: refactor update
